+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Scatter Plot: Added a `ConnectStyle` to enable step display mode (#3242) _Thanks @NoahSigl_
 * Polygon: `Plot.Add.Polygon()` now accepts generic type lists and arrays (#3244) _Thanks howhowone_23_
 * Demo: Added a draggable points window to show how to drag points of a scatter plot (#3248) _Thanks bologna_
+* Generate: Added `RandomNumber()` and `RandomNumbers()` overloads
 
 ## ScottPlot 4.1.71 (not yet on NuGet)
 * Assembly: All ScottPlot packages use the same strong name signing key (#3235, #3241) _Thanks @mlessmann_